+++ src/core/moves/SingleRandomMoveSchedule.h
@@ -27,7 +27,7 @@ namespace RevBayesCore {
         SingleRandomMoveSchedule*                             clone(void) const;
         double                                          getNumberMovesPerIteration(void) const;
         void                                            setNumberMovesPerIteration(double);
-        Move*                                           nextMove(size_t g);
+        Move*                                           nextMove(unsigned long g);
         
     private:
